Separate matrix class and corresponding unit tests
